requirements.txt
absl-py==0.6.1
astor==0.7.1
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
cleverhans==3.0.1
configparser==3.7.1
cycler==0.10.0
Cython==0.29.5
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
enum34==1.1.6
foolbox==1.8.0
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.2.0
gast==0.2.0
gitdb2==2.0.5
GitPython==2.1.11
grpcio==1.17.1
h5py==2.8.0
idna==2.8
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==2.2.3
mistune==0.8.4
mmdnn==0.2.5
mnist==0.2.2
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
nose==1.3.7
notebook==5.7.4
numpy==1.16.1
pandas==0.24.2
pandocfilters==1.4.2
pathlib2==2.3.3
pbr==5.1.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.5.0
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.3.1
pyparsing==2.3.0
python-dateutil==2.8.0
pytz==2018.7
PyYAML==5.1
pyzmq==17.1.2
qtconsole==4.4.3
randomgen==1.16.0
requests==2.21.0
scandir==1.9.0
scipy==1.2.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
smmap2==2.0.5
subprocess32==3.5.3
tensorboard==1.12.1
tensorflow-gpu==1.12.0
tensorflow-probability==0.5.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
torch==1.0.1.post2
torchvision==0.2.1
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
